# Small optimizations for `st_sfc()`