--- a/R/bbox.R
+++ b/R/bbox.R
@@ -10,14 +10,11 @@ bb_wrap = function(bb) {
 }
 
 bbox.Set = function(obj, ...) {
-	n_empty = sfc_count_empty(obj)
-	if (n_empty == length(obj))
-		return(NA_bbox_)
-	else if (n_empty == 0)
-		return(bb_wrap(CPL_get_bbox(obj, 0)))
-	
-	sel = vapply(obj, function(x) { length(x) && !all(is.na(x)) }, TRUE)
-	bb_wrap(CPL_get_bbox(unclass(obj)[sel], 0))
+	sel = !sfc_is_empty(obj)
+	if (! any(sel))
+		NA_bbox_
+	else
+		bb_wrap(CPL_get_bbox(unclass(obj)[sel], 0))
 }
 bbox.Mtrx = function(obj, ...) {
 	if (length(obj) == 0)