--- a/R/sfc.R
+++ b/R/sfc.R
@@ -48,12 +48,13 @@ st_sfc = function(..., crs = NA_crs_, precision = 0.0, check_ring_dir = FALSE, d
 
 	# check for NULLs:
 	a = attributes(lst)
-	is_null = vapply(lst, function(x) is.null(x) || isTRUE(is.na(x)), NA)
+	is_null = sfc_is_null(lst)
 	lst = unclass(lst)
 	lst = lst[! is_null]
 	attributes(lst) = a
 
-	sfg_classes = vapply(lst, class, rep(NA_character_, 3))
+	dims_and_types = sfc_unique_sfg_dims_and_types(lst)
+	
 	cls = if (length(lst) == 0) # empty set: no geometries read
 		c("sfc_GEOMETRY", "sfc")
 	else {
@@ -61,10 +62,10 @@ st_sfc = function(..., crs = NA_crs_, precision = 0.0, check_ring_dir = FALSE, d
 		single = if (!is.null(attr(lst, "single_type"))) # set by CPL_read_wkb:
 				attr(lst, "single_type")
 			else
-				length(unique(sfg_classes[2L,])) == 1L
+				length(dims_and_types[[2]]) == 1L
 		attr(lst, "single_type") = NULL # clean up
 		if (single)
-			c(paste0("sfc_", sfg_classes[2L, 1L]), "sfc")
+			c(paste0("sfc_", dims_and_types[[2]][1]), "sfc")
 		else
 			c("sfc_GEOMETRY", "sfc")    # the mix
 	}
@@ -74,14 +75,14 @@ st_sfc = function(..., crs = NA_crs_, precision = 0.0, check_ring_dir = FALSE, d
 			dim = if (length(lst) == 0) # we have no clue:
 					"XY"
 				else
-					sfg_classes[1L, 1L]
+					dims_and_types[[1]][1]
 		}
 		ret = vector("list", length(is_null))
 		ret[!is_null] = lst
 		ret[ is_null] = list(typed_empty(cls, nchar(dim), dim = dim))
 		attributes(ret) = attributes(lst)
 		lst = ret
-		sfg_classes = vapply(lst, class, rep(NA_character_, 3))
+		dims_and_types = sfc_unique_sfg_dims_and_types(lst)
 	}
 
 	# set class:
@@ -99,7 +100,7 @@ st_sfc = function(..., crs = NA_crs_, precision = 0.0, check_ring_dir = FALSE, d
 	# compute z_range, if dims permit and not set
 	zr = attr(lst, "z_range")
 	if (is.null(zr) || any(is.na(zr))) {
-		u <- unique(sfg_classes[1L,])
+		u <- dims_and_types[[1]]
 		if( "XYZM" %in% u ) {
 			attr(lst, "z_range") = compute_z_range(lst)
 			attr(lst, "m_range") = compute_m_range(lst)
@@ -125,20 +126,12 @@ st_sfc = function(..., crs = NA_crs_, precision = 0.0, check_ring_dir = FALSE, d
 
 	# set n_empty, check XY* is uniform:
 	if (is.null(attr(lst, "n_empty")) || any(is_null)) { # n_empty is set by CPL_read_wkb:
-		attr(lst, "n_empty") = sum(vapply(lst, sfg_is_empty, TRUE))
+		attr(lst, "n_empty") = sfc_count_empty(lst)
 # 		https://github.com/r-spatial/sf/issues/1592 :
 #		if (length(u <- unique(sfg_classes[1L,])) > 1)
 #			stop(paste("found multiple dimensions:", paste(u, collapse = " ")))
 	}
 	lst
-}
-
-sfg_is_empty = function(x) {
-	switch(class(x)[2],
-		POINT = any(!is.finite(x)),
-		MULTIPOINT = , LINESTRING = , CIRCULARSTRING = , CURVE = nrow(x) == 0,
-		length(x) == 0
-	)
 }
 
 #' @export
@@ -157,7 +150,7 @@ sfg_is_empty = function(x) {
 		value = list(value)
 	x = unclass(x) # becomes a list, but keeps attributes
 	ret = st_sfc(NextMethod())
-	structure(ret, n_empty = sum(vapply(ret, sfg_is_empty, TRUE)))
+	structure(ret, n_empty = sfc_count_empty(ret))
 }
 
 #' @export