--- a/src/RcppExports.cpp
+++ b/src/RcppExports.cpp
@@ -1088,6 +1088,39 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// sfc_is_null
+LogicalVector sfc_is_null(List sfc);
+RcppExport SEXP _sf_sfc_is_null(SEXP sfcSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type sfc(sfcSEXP);
+    rcpp_result_gen = Rcpp::wrap(sfc_is_null(sfc));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sfc_unique_sfg_dims_and_types
+List sfc_unique_sfg_dims_and_types(List sfc);
+RcppExport SEXP _sf_sfc_unique_sfg_dims_and_types(SEXP sfcSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type sfc(sfcSEXP);
+    rcpp_result_gen = Rcpp::wrap(sfc_unique_sfg_dims_and_types(sfc));
+    return rcpp_result_gen;
+END_RCPP
+}
+// sfc_count_empty
+int sfc_count_empty(List sfc);
+RcppExport SEXP _sf_sfc_count_empty(SEXP sfcSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< List >::type sfc(sfcSEXP);
+    rcpp_result_gen = Rcpp::wrap(sfc_count_empty(sfc));
+    return rcpp_result_gen;
+END_RCPP
+}
 // points_cpp
 List points_cpp(NumericMatrix pts, CharacterVector gdim);
 RcppExport SEXP _sf_points_cpp(SEXP ptsSEXP, SEXP gdimSEXP) {
@@ -1409,6 +1442,9 @@ static const R_CallMethodDef CallEntries[] = {
     {"_sf_CPL_proj_direct", (DL_FUNC) &_sf_CPL_proj_direct, 5},
     {"_sf_CPL_proj_info", (DL_FUNC) &_sf_CPL_proj_info, 1},
     {"_sf_CPL_xy2sfc", (DL_FUNC) &_sf_CPL_xy2sfc, 4},
+    {"_sf_sfc_is_null", (DL_FUNC) &_sf_sfc_is_null, 1},
+    {"_sf_sfc_unique_sfg_dims_and_types", (DL_FUNC) &_sf_sfc_unique_sfg_dims_and_types, 1},
+    {"_sf_sfc_count_empty", (DL_FUNC) &_sf_sfc_count_empty, 1},
     {"_sf_points_cpp", (DL_FUNC) &_sf_points_cpp, 2},
     {"_sf_CPL_signed_area", (DL_FUNC) &_sf_CPL_signed_area, 1},
     {"_sf_CPL_get_metadata", (DL_FUNC) &_sf_CPL_get_metadata, 3},