Following discussion on #1925, this PR eliminates `vapply()`s in most cases when constructing or subsetting, which mostly happens in `st_sfc()`. The most dramatic speedup is on points, where calculating the bounding box was very slow, but it looks like the speedup for polygons is about the same (about 8x faster, for both). 

I tried to only swap out the vapplys to avoid accidentally changing the structure of sfc objects! In #1925 @edzer experimented with a few ways to skip some of these checks, which may help even more. That said, I think this PR reduces the subset/construction overhead to the point where it's not noticeable compared to other things (like actual computations/reading from disk).

Current dev:

``` r
library(sf)
#> Linking to GEOS 3.10.1, GDAL 3.3.3, PROJ 8.1.0; sf_use_s2() is TRUE
n <- 1e6
points <- data.frame(id = seq_len(n), x = runif(n), y = runif(n))
points_sf <- st_as_sf(points, coords = c("x", "y"))
bench::mark(
  points_sf$geometry[],
  check = F,
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 points_sf$geometry[]    1.67s    1.76s     0.563     123MB     2.76
```

``` r
library(sf)
#> Linking to GEOS 3.10.1, GDAL 3.3.3, PROJ 8.1.0; sf_use_s2() is TRUE
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))
nc_big <- rep(nc$geometry, 1000)
bench::mark(nc_big[], check = FALSE)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 nc_big[]      333ms    350ms      2.85    54.8MB     9.99
```

After this PR:

``` r
library(sf)
#> Linking to GEOS 3.10.1, GDAL 3.3.3, PROJ 8.1.0; sf_use_s2() is TRUE
n <- 1e6
points <- data.frame(id = seq_len(n), x = runif(n), y = runif(n))
points_sf <- st_as_sf(points, coords = c("x", "y"))
bench::mark(
  points_sf$geometry[],
  check = F,
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 × 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 points_sf$geometry[]    169ms    247ms      3.25    61.3MB     4.87
```

``` r
library(sf)
#> Linking to GEOS 3.10.1, GDAL 3.3.3, PROJ 8.1.0; sf_use_s2() is TRUE
nc <- read_sf(system.file("shape/nc.shp", package = "sf"))
nc_big <- rep(nc$geometry, 1000)
bench::mark(nc_big[], check = FALSE)
#> # A tibble: 1 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 nc_big[]     41.5ms   41.5ms      24.1    48.2MB     145.
```